# stenella.openapi.json
{"openapi":"3.1.0","info":{"title":"Stenella","version":"0.1.0"},"paths":{"/api/v1/auth/access-token":{"post":{"tags":["auth"],"summary":"Login Access Token","description":"OAuth2 compatible token login, get an access token for future requests","operationId":"login_access_token_api_v1_auth_access_token_post","requestBody":{"content":{"application/x-www-form-urlencoded":{"schema":{"$ref":"#/components/schemas/Body_login_access_token_api_v1_auth_access_token_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"title":"Response Login Access Token Api V1 Auth Access Token Post"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/auth/magic-link":{"post":{"tags":["auth"],"summary":"Request Magic Link","description":"Request a magic link to be sent to the user's email.","operationId":"request_magic_link_api_v1_auth_magic_link_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/MagicLinkRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"title":"Response Request Magic Link Api V1 Auth Magic Link Post"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/auth/verify":{"get":{"tags":["auth"],"summary":"Verify Magic Link","description":"Verify a magic link token and log the user in.","operationId":"verify_magic_link_api_v1_auth_verify_get","parameters":[{"name":"email","in":"query","required":true,"schema":{"type":"string","title":"Email"}},{"name":"token","in":"query","required":true,"schema":{"type":"string","title":"Token"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"title":"Response Verify Magic Link Api V1 Auth Verify Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/users/":{"get":{"tags":["users"],"summary":"Read Users","description":"Retrieve users.","operationId":"read_users_api_v1_users__get","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"skip","in":"query","required":false,"schema":{"type":"integer","default":0,"title":"Skip"}},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","default":100,"title":"Limit"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/UserRead"},"title":"Response Read Users Api V1 Users  Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"post":{"tags":["users"],"summary":"Create User","description":"Create new user.","operationId":"create_user_api_v1_users__post","requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserCreate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserRead"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/users/me":{"get":{"tags":["users"],"summary":"Read User Me","description":"Get current user.","operationId":"read_user_me_api_v1_users_me_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserProfileOut"}}}}},"security":[{"OAuth2PasswordBearer":[]}]},"delete":{"tags":["users"],"summary":"Delete User Me","description":"Deactivate own user.","operationId":"delete_user_me_api_v1_users_me_delete","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"title":"Response Delete User Me Api V1 Users Me Delete"}}}}},"security":[{"OAuth2PasswordBearer":[]}]},"patch":{"tags":["users"],"summary":"Update User Me","description":"Update own user.","operationId":"update_user_me_api_v1_users_me_patch","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserUpdate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserProfileOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/v1/users/me/sessions/invalidate":{"post":{"tags":["users"],"summary":"Invalidate Own Sessions","description":"Invalidate all sessions for the current user (Global Logout).","operationId":"invalidate_own_sessions_api_v1_users_me_sessions_invalidate_post","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"title":"Response Invalidate Own Sessions Api V1 Users Me Sessions Invalidate Post"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/v1/users/pending-magic-links":{"get":{"tags":["users"],"summary":"Read Pending Magic Links","description":"Retrieve users with pending magic links.","operationId":"read_pending_magic_links_api_v1_users_pending_magic_links_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/UserWithMagicLink"},"type":"array","title":"Response Read Pending Magic Links Api V1 Users Pending Magic Links Get"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/v1/metrics/dau":{"get":{"tags":["metrics"],"summary":"Get Daily Active Users","description":"Get Total Daily Active Users (DAU) for today.","operationId":"get_daily_active_users_api_v1_metrics_dau_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/MetricCount"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/v1/metrics/mau":{"get":{"tags":["metrics"],"summary":"Get Monthly Active Users","description":"Get Total Monthly Active Users (MAU) for the rolling 30 days.","operationId":"get_monthly_active_users_api_v1_metrics_mau_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/MetricCount"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/v1/metrics/active-users":{"get":{"tags":["metrics"],"summary":"Get Active Users By Day","description":"Returns a dictionary grouping dates (YYYY-MM-DD) to a list of user_ids that were active on that date.","operationId":"get_active_users_by_day_api_v1_metrics_active_users_get","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"days","in":"query","required":false,"schema":{"type":"integer","default":30,"title":"Days"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"object","additionalProperties":{"type":"array","items":{"type":"integer"}},"title":"Response Get Active Users By Day Api V1 Metrics Active Users Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/metrics/signups":{"get":{"tags":["metrics"],"summary":"Get Daily Signups","description":"Returns the count of new user signups grouped by day for the last N days.","operationId":"get_daily_signups_api_v1_metrics_signups_get","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"days","in":"query","required":false,"schema":{"type":"integer","default":30,"title":"Days"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/DailyMetric"},"title":"Response Get Daily Signups Api V1 Metrics Signups Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/metrics/sessions":{"get":{"tags":["metrics"],"summary":"Get Session Metrics","description":"Returns metrics regarding active database sessions (if USE_DB_SESSIONS is true).","operationId":"get_session_metrics_api_v1_metrics_sessions_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SessionMetrics"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/v1/admin/users":{"post":{"tags":["admin"],"summary":"Create User","description":"Create a new user. Restricted to admin only.\nAllows specifying role and is_active status directly.","operationId":"create_user_api_v1_admin_users_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserCreate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserRead"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/v1/admin/users/active":{"get":{"tags":["admin"],"summary":"Read Active Users","description":"Retrieve all active users along with their session count and last active time.","operationId":"read_active_users_api_v1_admin_users_active_get","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"skip","in":"query","required":false,"schema":{"type":"integer","default":0,"title":"Skip"}},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","default":100,"title":"Limit"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/UserAdminView"},"title":"Response Read Active Users Api V1 Admin Users Active Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/admin/users/{user_ident}":{"get":{"tags":["admin"],"summary":"Read User","description":"Get a specific user by ID or Email.","operationId":"read_user_api_v1_admin_users__user_ident__get","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"user_ident","in":"path","required":true,"schema":{"type":"string","title":"User Ident"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserRead"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/admin/users/{user_id}":{"patch":{"tags":["admin"],"summary":"Update User","description":"Update a user. Update only if there is at least one field to update.","operationId":"update_user_api_v1_admin_users__user_id__patch","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"user_id","in":"path","required":true,"schema":{"type":"integer","title":"User Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserUpdateAdmin"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserRead"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["admin"],"summary":"Delete User","description":"Deactivate a user.","operationId":"delete_user_api_v1_admin_users__user_id__delete","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"user_id","in":"path","required":true,"schema":{"type":"integer","title":"User Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"title":"Response Delete User Api V1 Admin Users  User Id  Delete"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/admin/users/{user_id}/sessions/invalidate":{"post":{"tags":["admin"],"summary":"Invalidate User Sessions","description":"Invalidate all sessions for a specific user (Global Logout).","operationId":"invalidate_user_sessions_api_v1_admin_users__user_id__sessions_invalidate_post","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"user_id","in":"path","required":true,"schema":{"type":"integer","title":"User Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"title":"Response Invalidate User Sessions Api V1 Admin Users  User Id  Sessions Invalidate Post"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/":{"get":{"summary":"Root","operationId":"root__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/health":{"get":{"summary":"Health Check","operationId":"health_check_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"additionalProperties":{"type":"string"},"type":"object","title":"Response Health Check Health Get"}}}}}}}},"components":{"schemas":{"Body_login_access_token_api_v1_auth_access_token_post":{"properties":{"grant_type":{"anyOf":[{"type":"string","pattern":"^password$"},{"type":"null"}],"title":"Grant Type"},"username":{"type":"string","title":"Username"},"password":{"type":"string","format":"password","title":"Password"},"scope":{"type":"string","title":"Scope","default":""},"client_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Client Id"},"client_secret":{"anyOf":[{"type":"string"},{"type":"null"}],"format":"password","title":"Client Secret"}},"type":"object","required":["username","password"],"title":"Body_login_access_token_api_v1_auth_access_token_post"},"DailyMetric":{"properties":{"date":{"type":"string","title":"Date"},"count":{"type":"integer","title":"Count"}},"type":"object","required":["date","count"],"title":"DailyMetric"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"MagicLinkRequest":{"properties":{"email":{"type":"string","format":"email","title":"Email"}},"type":"object","required":["email"],"title":"MagicLinkRequest"},"MetricCount":{"properties":{"count":{"type":"integer","title":"Count"}},"type":"object","required":["count"],"title":"MetricCount"},"SessionMetrics":{"properties":{"total_active_sessions":{"type":"integer","title":"Total Active Sessions"},"unique_users_logged_in":{"type":"integer","title":"Unique Users Logged In"}},"type":"object","required":["total_active_sessions","unique_users_logged_in"],"title":"SessionMetrics"},"UserAdminView":{"properties":{"email":{"type":"string","title":"Email"},"display_name":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Display Name"},"role":{"type":"string","title":"Role","default":"user"},"is_active":{"type":"boolean","title":"Is Active","default":true},"id":{"type":"integer","title":"Id"},"created_at":{"type":"string","format":"date-time","title":"Created At"},"session_count":{"type":"integer","title":"Session Count","default":0},"last_active_time":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Last Active Time"}},"type":"object","required":["email","id","created_at"],"title":"UserAdminView"},"UserCreate":{"properties":{"email":{"type":"string","title":"Email"},"display_name":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Display Name"},"role":{"type":"string","title":"Role","default":"user"},"is_active":{"type":"boolean","title":"Is Active","default":true},"password":{"type":"string","title":"Password"}},"type":"object","required":["email","password"],"title":"UserCreate"},"UserProfileOut":{"properties":{"email":{"type":"string","title":"Email"},"display_name":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Display Name"},"role":{"type":"string","title":"Role","default":"user"},"is_active":{"type":"boolean","title":"Is Active","default":true},"id":{"type":"integer","title":"Id"},"created_at":{"type":"string","format":"date-time","title":"Created At"},"has_password":{"type":"boolean","title":"Has Password"}},"type":"object","required":["email","id","created_at","has_password"],"title":"UserProfileOut"},"UserRead":{"properties":{"email":{"type":"string","title":"Email"},"display_name":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Display Name"},"role":{"type":"string","title":"Role","default":"user"},"is_active":{"type":"boolean","title":"Is Active","default":true},"id":{"type":"integer","title":"Id"},"created_at":{"type":"string","format":"date-time","title":"Created At"}},"type":"object","required":["email","id","created_at"],"title":"UserRead"},"UserUpdate":{"properties":{"display_name":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Display Name"},"password":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Password"}},"type":"object","title":"UserUpdate"},"UserUpdateAdmin":{"properties":{"display_name":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Display Name"},"password":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Password"},"role":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Role"},"is_active":{"anyOf":[{"type":"boolean"},{"type":"null"}],"title":"Is Active"}},"type":"object","title":"UserUpdateAdmin"},"UserWithMagicLink":{"properties":{"email":{"type":"string","title":"Email"},"display_name":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Display Name"},"role":{"type":"string","title":"Role","default":"user"},"is_active":{"type":"boolean","title":"Is Active","default":true},"id":{"type":"integer","title":"Id"},"created_at":{"type":"string","format":"date-time","title":"Created At"},"magic_token":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Magic Token"},"magic_token_expires_at":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Magic Token Expires At"}},"type":"object","required":["email","id","created_at"],"title":"UserWithMagicLink"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"},"input":{"title":"Input"},"ctx":{"type":"object","title":"Context"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}},"securitySchemes":{"OAuth2PasswordBearer":{"type":"oauth2","flows":{"password":{"scopes":{},"tokenUrl":"/api/v1/auth/access-token"}}}}}}